+++ Tarteeb.Api/Startup.cs
@@ -27,6 +27,7 @@ using Tarteeb.Api.Services.Foundations.Tickets;
 using Tarteeb.Api.Services.Foundations.Times;
 using Tarteeb.Api.Services.Foundations.Users;
 using Tarteeb.Api.Services.Orchestrations;
+using Tarteeb.Api.Services.Orchestrations.UserSecurities;
 using Tarteeb.Api.Services.Processings.Users;
 
 namespace Tarteeb.Api
